Feature/refactoring

+++ metastats-infrastructure/src/main/groovy/life/qbic/metastats/request/OpenBisSearch.groovy
@@ -27,7 +27,6 @@ import org.apache.logging.log4j.Logger
 
 class OpenBisSearch implements DatabaseGateway {
 
-
     Project project
     IApplicationServerApi v3
     IDataStoreServerApi dss
@@ -37,8 +36,12 @@ class OpenBisSearch implements DatabaseGateway {
     OpenBisParser parser = new OpenBisParser()
 
 
-    private static final Logger LOG = LogManager.getLogger(OpenBisSearch.class);
+    private static final Logger LOG = LogManager.getLogger(OpenBisSearch.class)
 
+    /**
+     * Creates an openbis search for a given session
+     * @param session used to query openbis
+     */
     OpenBisSearch(OpenBisSession session) {
         this.session = session
         this.v3 = session.v3
@@ -69,6 +72,12 @@ class OpenBisSearch implements DatabaseGateway {
         fetchOptions.withExperimentsUsing(experiment)
 
         SearchResult<Project> resProject = v3.searchProjects(sessionToken, criteria, fetchOptions)
+
+        if (resProject.getObjects().size() == 0) {
+            LOG.warn "Project $projectCode not found"
+            System.exit(-1)
+        }
+
         project = resProject.getObjects().get(0)
     }
 
@@ -97,6 +106,10 @@ class OpenBisSearch implements DatabaseGateway {
         session.logout()
     }
 
+    /**
+     * Retrieves all samples on the level of biological entities (Q_BIOLOGICAL_ENTITY)
+     * @return a list of metastats samples
+     */
     List<MetaStatsSample> fetchBiologicalEntity() {
         List prepSamples = []
         List<Experiment> experiments = project.experiments
@@ -104,7 +117,7 @@ class OpenBisSearch implements DatabaseGateway {
 
         experiments.each { exp ->
             //only do it for the Experiment with Biological Samples since from here all other samples can be found
-            if (exp.code == project.code + "E2" && exp.getSamples() != null) {
+            if (exp.type.code == "Q_EXPERIMENTAL_DESIGN" && exp.getSamples() != null) {
                 //todo manchmal gibts e2 2x --> only consider exp with samples
                 LOG.info "found $exp.type.code"
 
@@ -123,31 +136,44 @@ class OpenBisSearch implements DatabaseGateway {
         return prepSamples
     }
 
-    def translateSampleVocabulary(List<MetaStatsSample> samples) {
+    /**
+     * Translates the vocabulary of a given list of samples
+     * @param samples containing properties that require translation e.g Q_NCBI_ORGANISM
+     */
+    void translateSampleVocabulary(List<MetaStatsSample> samples) {
         samples.each { sample ->
-            sample.properties.each { key, value ->
+            sample.sampleProperties.each { key, value ->
                 if (key == "Q_NCBI_ORGANISM" || key == "Q_PRIMARY_TISSUE") {
                     String vocabulary = fetchVocabulary(value)
                     //overwrite old key
-                    sample.properties.put(key, vocabulary)
+                    sample.sampleProperties.put(key, vocabulary)
                 }
             }
         }
     }
 
-    def translateExperimentVocabulary(List<MetaStatsExperiment> experiments) {
+    /**
+     * Translates the Experiment vocabulary
+     * @param experiments containing properties that require translation e.g Q_SEQUENCER_DEVICE
+     */
+    void translateExperimentVocabulary(List<MetaStatsExperiment> experiments) {
         experiments.each { experiment ->
-            experiment.properties.each { key, value ->
+            experiment.experimentProperties.each { key, value ->
                 if (key == "Q_SEQUENCER_DEVICE") {
                     String vocabulary = fetchVocabulary(value)
                     //overwrite old key
-                    experiment.properties.put(key, vocabulary)
+                    experiment.experimentProperties.put(key, vocabulary)
                 }
             }
         }
     }
 
-    def fetchVocabulary(String code) {
+    /**
+     * Fetches the vocabulary that needs to be translated from openbis
+     * @param code
+     * @return translated term
+     */
+    String fetchVocabulary(String code) {
         VocabularyTermSearchCriteria vocabularyTermSearchCriteria = new VocabularyTermSearchCriteria()
         vocabularyTermSearchCriteria.withCode().thatEquals(code)
 
@@ -158,17 +184,30 @@ class OpenBisSearch implements DatabaseGateway {
         return term
     }
 
-    def addFile(List<MetaStatsSample> prepSamples) {
+    /**
+     * Fetches and adds files to a given list of samples
+     * @param prepSamples to which corresponding files are added
+     */
+    void addFile(List<MetaStatsSample> prepSamples) {
         prepSamples.each { sample ->
-            HashMap files = fetchDataSets(sample.code, "fastq")
-            sample.properties << files
+            HashMap files = fetchDataSets(sample.sampleCode, "fastq")
+            sample.sampleProperties << files
         }
     }
 
+    /**
+     * Fetches datasets of a given filetype for a sample specified by its code
+     * @param sampleCode specifying the sample for which the DS shall be fetched
+     * @param fileType specifying which files are searched
+     * @return a maps of found datasets with their openbis type
+     */
     HashMap<String, String> fetchDataSets(String sampleCode, String fileType) {
         LOG.info "fetch DataSet for $sampleCode"
 
         HashMap allDataSets = new HashMap()
+
+        StringBuilder dataFiles = new StringBuilder("")
+        String datasetType = ""
 
         findAllDatasetsRecursive(sampleCode).each { dataSet ->
             DataSetFileSearchCriteria criteria = new DataSetFileSearchCriteria()
@@ -178,22 +217,32 @@ class OpenBisSearch implements DatabaseGateway {
             SearchResult<DataSetFile> result = dss.searchFiles(sessionToken, criteria, new DataSetFileFetchOptions())
             List<DataSetFile> files = result.getObjects()
 
-            StringBuilder dataFiles = new StringBuilder()
-
             for (DataSetFile file : files) {
                 if (file.getPermId().toString().contains("." + fileType)
                         && !file.getPermId().toString().contains(".sha256sum")
-                        && !file.getPermId().toString().contains("origlabfilename")) {
+                        && !file.getPermId().toString().contains("origlabfilename")
+                        && !file.getPermId().toString().contains(".csv")
+                        && !file.getPermId().toString().contains(".txt")) {
                     String[] path = file.getPermId().toString().split("/")
-                    dataFiles << path[path.size() - 1] + ", "
+                    if (!dataFiles.contains(path[path.size() - 1])) dataFiles << path[path.size() - 1] + ", "
+                    datasetType = dataSet.type.code
                 }
             }
-            dataFiles.delete(dataFiles.length() - 2, dataFiles.length())
-            allDataSets.put(dataSet.type.code, dataFiles)
         }
+
+        if (dataFiles.toString() != "") {
+            dataFiles.delete(dataFiles.length() - 2, dataFiles.length())
+            allDataSets.put(datasetType, dataFiles)
+        }
+
         return allDataSets
     }
 
+    /**
+     * Searches for all related datasets of a sample with a recursion
+     * @param sampleId specifying the sample for which DS are searched
+     * @return list of found datasets
+     */
     List<DataSet> findAllDatasetsRecursive(final String sampleId) {
         SampleSearchCriteria criteria = new SampleSearchCriteria()
         criteria.withCode().thatEquals(sampleId)
@@ -219,6 +268,11 @@ class OpenBisSearch implements DatabaseGateway {
         return foundDatasets
     }
 
+    /**
+     * Fetches descendant datstes of a sample
+     * @param sample
+     * @return list of descendant datasets
+     */
     private static List<DataSet> fetchDescendantDatasets(final Sample sample) {
         List<DataSet> foundSets = new ArrayList<>()
